// client/src/Redux/Constants/PlaneConstants.ts
export enum PlanesList {
    PLANE_LIST_REQUEST = "PLANE_LIST_REQUEST",
    PLANE_LIST_SUCCESS = "PLANE_LIST_SUCCESS",
    PLANE_LIST_FAIL = "PLANE_LIST_FAIL",
}

export enum PlaneDetails {
    PLANE_DETAILS_REQUEST = "PLANE_DETAILS_REQUEST",
    PLANE_DETAILS_SUCCESS = "PLANE_DETAILS_SUCCESS",
    PLANE_DETAILS_FAIL = "PLANE_DETAILS_FAIL",
}

export enum PlaneCreate {
    PLANE_CREATE_REQUEST = "PLANE_CREATE_REQUEST",
    PLANE_CREATE_SUCCESS = "PLANE_CREATE_SUCCESS",
    PLANE_CREATE_FAIL = "PLANE_CREATE_FAIL",
    PLANE_CREATE_RESET = "PLANE_CREATE_RESET"
}

export enum PlaneDelete {
    PLANE_DELETE_REQUEST = "PLANE_DELETE_REQUEST",
    PLANE_DELETE_SUCCESS = "PLANE_DELETE_SUCCESS",
    PLANE_DELETE_FAIL = "PLANE_DELETE_FAIL",
}